requirements.txt
aiofiles==0.5.0
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cymem==2.0.3
dataclasses==0.6
fastapi==0.44.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
ipython==7.15.0
ipython-genutils==0.2.0
murmurhash==1.0.2
numpy==1.18.4
pandas==1.0.4
peewee==3.13.3
plac==1.1.3
preshed==3.0.2
progress==1.5
psycopg2-binary==2.8.6
pydantic==1.5.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
spacy==2.3
sqlalchemy==1.3.19
srsly==1.0.2
starlette==0.12.9
thinc==7.4.1
toolz==0.10.0
tqdm==4.46.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wasabi==0.6.0
websockets==8.1
xlrd==1.2.0
zipp==3.1.0
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz#egg=de_core_news_sm
https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-2.3.0/de_core_news_lg-2.3.0.tar.gz#egg=de_core_news_lg
